rename: 정적 파일에 대한 핸들러도 서블릿으로 취급
- StaticResourceRequestHandler -> DefaultServlet
- 프로덕션 코드의 XXXHandler -> XXXServlet
