+++ tomcat/src/main/java/nextstep/jwp/servlet/LoginRequestServlet.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.handler;
+package nextstep.jwp.servlet;
 
 import java.io.IOException;
 import javassist.NotFoundException;
@@ -8,21 +8,21 @@ import org.apache.catalina.core.servlet.HttpServlet;
 import org.apache.catalina.core.servlet.HttpServletResponse;
 import org.apache.coyote.http11.request.Request;
 
-public class RegisterRequestHandler extends HttpServlet {
+public class LoginRequestServlet extends HttpServlet {
 
-    public RegisterRequestHandler() {
-        super("/register");
+    public LoginRequestServlet() {
+        super("/login");
     }
 
     @Override
     protected void doGet(final Request request, final HttpServletResponse response)
             throws NotFoundException, IOException {
-        response.set(ViewController.register(request));
+        response.set(ViewController.login(request));
     }
 
     @Override
     protected void doPost(final Request request, final HttpServletResponse response) {
-        response.set(UserController.register(request));
+        response.set(UserController.login(request));
     }
 
 }
